{{cookiecutter.ci_project_name}}/requirements.freeze.txt
aniso8601==7.0.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
autoflake==2.0.0
awesome-slugify==1.6.5
backcall==0.2.0
black==23.1a1
blessed==1.19.1
boto3==1.26.59
botocore==1.29.59
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
configparser==5.3.0
coverage==7.1.0
cssselect==1.2.0
cssutils==2.6.0
decorator==5.1.1
distlib==0.3.6
Django==3.2.16
django-anymail==9.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-fsm==2.8.1
django-graphql-jwt==0.3.0
django-memoize==2.3.1
django-minio-storage==0.3.10
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
django-storages==1.13.2
django-stubs==1.14.0
django-stubs-ext==0.7.0
django-webpack-loader==1.8.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
execnet==1.9.0
executing==1.2.0
factory-boy==3.2.1
Faker==16.6.1
filelock==3.9.0
flake8==6.0.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
graphene==2.1.9
graphene-django==2.15.0
graphene-validator==0.0.1
graphql-core==2.3.2
graphql-relay==2.0.1
identify==2.5.16
idna==3.4
iniconfig==2.0.0
ipython==8.9.0
isort==6.0.0b2
jedi==0.18.2
jmespath==1.0.1
kubernetes==26.1.0a1
lxml==4.9.2
matplotlib-inline==0.1.6
mccabe==0.7.0
minio==6.0.2
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
phonenumbers==8.13.5
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.1
premailer==3.10.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pycodestyle==2.10.0
pycurl==7.45.2
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==1.7.1
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-xdist==3.1.0
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2022.7.1
PyYAML==6.0
redis==3.5.3
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
Rx==1.6.3
s3transfer==0.6.0
sentry-sdk==1.14.0
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
tomli==2.0.1
traitlets==5.8.1
types-python-dateutil==2.8.19.6
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.3
types-requests==2.28.11.8
types-urllib3==1.26.25.4
typing_extensions==4.4.0
Unidecode==0.4.21
uritemplate==4.1.1
urllib3==1.26.14
uWSGI==2.0.21
virtualenv==20.17.1
wcwidth==0.2.6
websocket-client==1.5.0
whitenoise==6.3.0
